requirements.txt
matplotlib
librosa==0.9.1
pytorch_lightning==1.9.0
pandas==1.1.5
numpy==1.21.0
tqdm==4.63.1
wordcloud==1.8.1
einops==0.8.0
scikit_learn==0.24.2
moviepy==1.0.3
Cython==0.29.28
huggingface_hub==0.23.2
pycocotools
packaging==21.3
torchvision==0.13.1+cu113
SPARQLWrapper==2.0.0
Pillow==8.4.0
opencv_python==4.6.0
jieba==0.42.1
transformers==4.41.2
dgl_cu111==0.6.1
torch==1.12.1+cu113
torchmetrics==1.4.0.post0
requests==2.27.1
